refactor: move uniq & isBsonType to a new helper file
