+++ packages/repository/src/relations/belongs-to/belongs-to.inclusion-resolver.ts
@@ -3,11 +3,11 @@
 // This file is licensed under the MIT License.
 // License text available at https://opensource.org/licenses/MIT
 
-import * as assert from 'assert';
 import {AnyObject} from '../../common-types';
 import {Entity} from '../../model';
 import {Inclusion} from '../../query';
 import {EntityCrudRepository} from '../../repositories/repository';
+import {uniq} from '../relation.helpers';
 import {
   BelongsToDefinition,
   Getter,
@@ -74,44 +74,5 @@ export class BelongsToInclusionResolver<
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       source[sourceProp] = targets[0] as any;
     }
-  }
-}
-
-// TODO(bajtos) move this to repository-level "helpers" or "utils" file
-// and add test coverage!!
-
-/**
- * Dedupe an array
- * @param {Array} input an array
- * @returns {Array} an array with unique items
- */
-function uniq<T>(input: T[]): T[] {
-  const uniqArray: T[] = [];
-  if (!input) {
-    return uniqArray;
-  }
-  assert(Array.isArray(input), 'array argument is required');
-
-  const comparableA = input.map(item =>
-    isBsonType(item) ? item.toString() : item,
-  );
-  for (let i = 0, n = comparableA.length; i < n; i++) {
-    if (comparableA.indexOf(comparableA[i]) === i) {
-      uniqArray.push(input[i]);
-    }
-  }
-  return uniqArray;
-}
-
-// TODO(bajtos) move this to repository-level "helpers" or "utils" file
-// and add test coverage!!
-function isBsonType(value: unknown): value is object {
-  if (typeof value !== 'object' || !value) return false;
-
-  // bson@1.x stores _bsontype on ObjectID instance, bson@4.x on prototype
-  return check(value) || check(value.constructor.prototype);
-
-  function check(target: unknown) {
-    return Object.prototype.hasOwnProperty.call(target, '_bsontype');
   }
 }
