+++ packages/repository/src/relations/relation.helpers.ts
@@ -1,0 +1,43 @@
+// Copyright IBM Corp. 2019. All Rights Reserved.
+// Node module: @loopback/repository
+// This file is licensed under the MIT License.
+// License text available at https://opensource.org/licenses/MIT
+
+import * as assert from 'assert';
+
+// TODO(bajtos) add test coverage
+
+/**
+ * Dedupe an array
+ * @param {Array} input an array
+ * @returns {Array} an array with unique items
+ */
+export function uniq<T>(input: T[]): T[] {
+  const uniqArray: T[] = [];
+  if (!input) {
+    return uniqArray;
+  }
+  assert(Array.isArray(input), 'array argument is required');
+
+  const comparableA = input.map(item =>
+    isBsonType(item) ? item.toString() : item,
+  );
+  for (let i = 0, n = comparableA.length; i < n; i++) {
+    if (comparableA.indexOf(comparableA[i]) === i) {
+      uniqArray.push(input[i]);
+    }
+  }
+  return uniqArray;
+}
+
+// TODO(bajtos) add test coverage
+export function isBsonType(value: unknown): value is object {
+  if (typeof value !== 'object' || !value) return false;
+
+  // bson@1.x stores _bsontype on ObjectID instance, bson@4.x on prototype
+  return check(value) || check(value.constructor.prototype);
+
+  function check(target: unknown) {
+    return Object.prototype.hasOwnProperty.call(target, '_bsontype');
+  }
+}
